+++ wojilu.Apps/Forum/Interface/IForumPickService.cs
@@ -12,12 +12,12 @@ namespace wojilu.Apps.Forum.Interface {
         ForumPick GetEditTopic( int appId, int topicId );
         ForumPick GetPinPostByIndex( int appId, int index );
 
-        Result AddPinPost( int appId, int index, string title, string link, string summary );
+        Result AddPinPost(long appId, int index, string title, string link, string summary);
         Result AddPinTopic( ForumTopic topic, int index );
 
-        void DeletePinPost( int appId, int index );
-        void DeleteTopic( int appId, int topicId );
-        void RestoreTopic( int appId, int topicId );
+        void DeletePinPost(long appId, int index);
+        void DeleteTopic(long appId, long topicId);
+        void RestoreTopic(long appId, long topicId);
 
         void EditPinPost( ForumPick x, int index, string title, string link, string summary );
         void EditTopic( int appId, int topicId, string title, string link, string summary );
