+++ wojilu.Core/Common/Feeds/Domain/Feed.cs
@@ -19,7 +19,7 @@ namespace wojilu.Common.Feeds.Domain {
 
         public User Creator { get; set; }
         public String DataType { get; set; }
-        public int DataId { get; set; }
+        public long DataId { get; set; }
 
         public int Replies { get; set; }
