+++ wojilu.Core/Common/Tag/DataTagShip.cs
@@ -10,7 +10,7 @@ namespace wojilu.Common.Tags {
     [Serializable]
     public class DataTagShip : ObjectBase<DataTagShip> {
 
-        public int DataId { get; set; }
+        public long DataId { get; set; }
         public Tag Tag { get; set; }
         public String TypeFullName { get; set; }
         public DateTime Created { get; set; }
